data refactoring

+++ Drivers/LSM9DS1-Driver/LSM9DS1_Driver.c
@@ -783,9 +783,9 @@ LSM9DS1_XLG_READ LSM9DS1_Read_XLG(MFX_input_t *data_in, int samples)
 
 	     }
 	     /* Format the data. */
-	     data_in->gyro[0] = ( int32_t ) ((pData[0] * sensitivityG )/divisor);
-	     data_in->gyro[1] = ( int32_t ) ((pData[1] * sensitivityG )/divisor);
-	     data_in->gyro[2] = ( int32_t )(((pData[2] * sensitivityG )/divisor));
+	     data_in->gyro[0] =  (pData[0] * sensitivityG )/divisor;
+	     data_in->gyro[1] =  (pData[1] * sensitivityG )/divisor;
+	     data_in->gyro[2] = (pData[2] * sensitivityG )/divisor;
 
 	     regadd=LSM9DS1_OUT_X_L_XL;
 	     //read Accelerometer output
@@ -808,9 +808,9 @@ LSM9DS1_XLG_READ LSM9DS1_Read_XLG(MFX_input_t *data_in, int samples)
 
 	   	 }
 	   	     /* Format the data. */
-	   	 	 data_in->acc[0] = ( int32_t ) (pData[3] * sensitivityXL)/divisor;
-	   	 	 data_in->acc[1] = ( int32_t ) (pData[4] * sensitivityXL)/divisor;
-	   	 	 data_in->acc[2] = ( int32_t )(pData[5] * sensitivityXL)/divisor;
+	   	 	 data_in->acc[0] =  (pData[3] * sensitivityXL)/divisor;
+	   	 	 data_in->acc[1] =  (pData[4] * sensitivityXL)/divisor;
+	   	 	 data_in->acc[2] = (pData[5] * sensitivityXL)/divisor;
 	   }
 	    return 1;
 }
